+++ inputs.tf
@@ -65,7 +65,12 @@ variable "availability_set_id" {
 }
 
 variable "diagnostics_storage_account_name" {
-  description = "Storage account name to store vm boot diagnostic"
+  description = "Name of the Storage Account in which store vm diagnostics"
+  type        = string
+}
+
+variable "diagnostics_storage_account_key" {
+  description = "Access key of the Storage Account in which store vm diagnostics"
   type        = string
 }