+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Breaking
 
   * AZ-94 Terraform 0.12 / HCL2 format
+  
+Added
+
+  * AZ-113 Enable diagnostics settings 
 
 # v0.1.0 - 2019-07-12